requirements.txt
discord.py
iniparser2